refactor(storybook): move storybook to apps
