+++ apps/storybook/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@cowprotocol/storybook",
+  "version": "0.0.0",
+  "private": true,
+  "devDependencies": {
+    "@storybook/addon-a11y": "^10.3.1",
+    "@storybook/addon-docs": "^10.3.1",
+    "@storybook/react": "10.3.1",
+    "@storybook/react-vite": "^10.3.1",
+    "@types/react": "19.1.3",
+    "@types/react-dom": "19.1.3",
+    "@types/styled-components": "5.1.34",
+    "react": "19.1.2",
+    "react-dom": "19.1.2",
+    "storybook": "^10.3.1",
+    "styled-components": "5.3.11"
+  }
+}
