+++ apps/storybook/project.json
@@ -1,0 +1,15 @@
+{
+  "name": "storybook",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "projectType": "application",
+  "sourceRoot": "apps/storybook/src",
+  "targets": {
+    "storybook": {
+      "command": "STORYBOOK_DISABLE_TELEMETRY=1 HOME=.cache/storybook-home storybook dev -p 6006 -c apps/storybook/src"
+    },
+    "build": {
+      "command": "STORYBOOK_DISABLE_TELEMETRY=1 HOME=.cache/storybook-home storybook build --output-dir build/storybook -c apps/storybook/src",
+      "outputs": ["{workspaceRoot}/build/storybook"]
+    }
+  }
+}
