+++ package.json
@@ -17,8 +17,8 @@
     "start:cosmos": "nx run cowswap-frontend:cosmos:run",
     "start:cosmos:explorer": "nx run explorer:cosmos:run",
     "start:explorer": "cross-env NODE_ENV=development nx run explorer:serve",
-    "storybook": "STORYBOOK_DISABLE_TELEMETRY=1 HOME=.cache/storybook-home storybook dev -p 6006",
-    "storybook:build": "STORYBOOK_DISABLE_TELEMETRY=1 HOME=.cache/storybook-home storybook build --output-dir build/storybook",
+    "storybook": "nx run storybook:storybook",
+    "storybook:build": "nx run storybook:build",
     "start:cowfi": "cross-env PORT=3001 nx run cow-fi:dev",
     "start:sdk-tools": "nx run sdk-tools:serve",
     "build": "pnpm run build:cowswap",
@@ -137,9 +137,6 @@
     "@nx/vite": "22.4.2",
     "@nx/workspace": "22.4.2",
     "@sentry/vite-plugin": "^4.4.0",
-    "@storybook/addon-a11y": "^10.3.1",
-    "@storybook/react": "10.3.1",
-    "@storybook/react-vite": "^10.3.1",
     "@swc-node/register": "~1.11.1",
     "@swc/cli": "~0.7.5",
     "@swc/core": "~1.13.5",
@@ -150,10 +147,7 @@
     "@types/d3": "^7.4.0",
     "@types/jest": "^30.0.0",
     "@types/node": "24.7.1",
-    "@types/react": "19.1.3",
-    "@types/react-dom": "19.1.3",
     "@types/react-is": "19.0.0",
-    "@types/styled-components": "5.1.34",
     "@types/wcag-contrast": "^3.0.0",
     "@typescript-eslint/eslint-plugin": "^8.4.0",
     "@typescript-eslint/parser": "^8.4.0",
@@ -195,12 +189,8 @@
     "nx-cloud": "19.1.0",
     "patch-package": "^8.0.0",
     "prettier": "~3.6.2",
-    "react": "19.1.2",
     "react-cosmos-plugin-vite": "^7.0.0",
-    "react-dom": "19.1.2",
     "rollup-plugin-visualizer": "^6.0.4",
-    "storybook": "^10.3.1",
-    "styled-components": "5.3.11",
     "ts-jest": "^29.1.1",
     "ts-node": "10.9.1",
     "typechain": "^8.3.2",
@@ -214,7 +204,6 @@
     "vite-plugin-pwa": "^1.0.0",
     "vite-plugin-svgr": "^4.2.0",
     "vite-tsconfig-paths": "~5.1.4",
-    "vitest": "4.0.9",
-    "@storybook/addon-docs": "^10.3.1"
+    "vitest": "4.0.9"
   }
 }
